requirements.txt
Django==3.2
psycopg2>=2.8
djangorestframework==3.13.1
Pillow==9.0.1
PyJWT==2.3.0